Rename counters.ts -> EphemeralCounterSource.ts, add tests to get full coverage
